ai-service/requirements.txt
opencv-python-headless==4.10.0.84
numpy==2.1.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
groq==0.11.0
python-dotenv==1.0.1
requests==2.32.3
